Msutils.String (mscmp_syst_utils_string v0.1.0)

This is a set of utilities for working with strings which is generally useful across components.

Functions

get_random_string(string_length, tokens \\ :alphanum)

@spec get_random_string(pos_integer(), MscmpSystUtilsString.Types.tokens()) ::
  String.t()

Generates a random string drawn from a specified list of characters.

Parameters

string_length - the number of characters in the returned string.

tokens - this optional parameter may either be a charlist() including the desired characters from which to randomly select characters for the string or the parameter may be an atom which designates a predefined character list. The available predefined character lists are:

:alphanum - will return values from the set 0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ. This is the default value.

:mixed_alphanum - will return values from the set 0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz

b32e - will return values from the set 0123456789ABCDEFGHIJKLMNOPQRSTUV. This equates to the character set used by Elixir's Integer.to_string(x, 32).

b32c - will return values from the set 0123456789ABCDEFGHJKMNPQRSTVWXYZ. This is the Base32 character set compatible with Douglas Crockford's Base 32 (https://www.crockford.com/base32.html).